server_python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosqlite==0.20.0
pydantic[email]==2.9.0
python-dotenv==1.0.1
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
